+++ packages/eslint-config/index.js
@@ -39,7 +39,6 @@ module.exports = {
     requirejs: false
   },
   plugins: ['jest', 'unused-imports'],
-  ignorePatterns: ['packages/web-integration-oc10/js'],
   overrides: [
     {
       files: ['**/*.vue'],