+++ packages/web-app-files/src/search/sdk/list.ts
@@ -37,6 +37,8 @@ export default class List implements SearchList {
   }
 
   async search(term: string): Promise<SearchResult> {
+    const useSpacesEndpoint = this.store.getters.capabilities?.spaces?.enabled === true
+
     if (!term) {
       return {
         totalResults: null,
@@ -47,7 +49,8 @@ export default class List implements SearchList {
     const { range, results } = await this.clientService.owncloudSdk.files.search(
       term,
       searchLimit,
-      DavProperties.Default
+      DavProperties.Default,
+      useSpacesEndpoint
     )
 
     return {