+++ packages/web-app-importer/tests/unit/extensions.spec.ts
@@ -55,6 +55,22 @@ describe('useFileActionsImport', () => {
         }
       })
     })
+    it('is false when no supported clouds are given', () => {
+      getWrapper({
+        currentFolder: mock<Resource>({ canUpload: () => true }),
+        setup: () => {
+          const action = unref(
+            extensions({
+              applicationConfig: {
+                companionUrl: 'companionUrl',
+                supportedClouds: []
+              }
+            })
+          )[0].action
+          expect(action.isEnabled()).toBeFalsy()
+        }
+      })
+    })
     it('is true on generic space view when write access is given', () => {
       getWrapper({
         currentFolder: mock<Resource>({ canUpload: () => true }),
@@ -89,6 +105,27 @@ describe('useFileActionsImport', () => {
         setup: () => {
           const action = getAction({ applicationConfig: { companionUrl: 'companionUrl' } })
           expect(action.isDisabled()).toBeFalsy()
+        }
+      })
+    })
+  })
+
+  describe('handler', () => {
+    it.each([
+      { supportedClouds: undefined, calls: 4 },
+      { supportedClouds: ['OneDrive'], calls: 2 },
+      { supportedClouds: ['GoogleDrive'], calls: 2 },
+      { supportedClouds: ['WebdavPublicLink'], calls: 2 }
+    ])('should only add supported clouds as uppy plugin', ({ supportedClouds, calls }) => {
+      const uppyService = mockDeep<UppyService>()
+      getWrapper({
+        uppyService,
+        setup: async () => {
+          const action = getAction({
+            applicationConfig: { companionUrl: 'companionUrl', supportedClouds }
+          })
+          await action.handler()
+          expect(uppyService.addPlugin).toHaveBeenCalledTimes(calls)
         }
       })
     })