+++ packages/web-runtime/tests/unit/components/Topbar/Notifications.spec.ts
@@ -24,6 +24,8 @@ const selectors = {
   notificationActions: '.oc-notifications-actions'
 }
 
+jest.mock('web-pkg/src/composables/sse/useServerSentEvents')
+
 describe('Notification component', () => {
   it('renders the notification bell and no notifications if there are none', () => {
     const { wrapper } = getWrapper()
@@ -34,6 +36,7 @@ describe('Notification component', () => {
   it('renders a set of notifications', async () => {
     const notifications = [mock<Notification>({ messageRich: undefined })]
     const { wrapper } = getWrapper({ notifications })
+    await wrapper.vm.fetchNotificationsTask.perform()
     await wrapper.vm.fetchNotificationsTask.last
     expect(wrapper.find(selectors.noNewNotifications).exists()).toBeFalsy()
     expect(wrapper.findAll(selectors.notificationItem).length).toBe(notifications.length)
@@ -41,6 +44,7 @@ describe('Notification component', () => {
   it('renders the loading state', async () => {
     const notifications = [mock<Notification>({ messageRich: undefined })]
     const { wrapper } = getWrapper({ notifications })
+    await wrapper.vm.fetchNotificationsTask.perform()
     await wrapper.vm.fetchNotificationsTask.last
     wrapper.vm.loading = true
     await wrapper.vm.$nextTick()
@@ -49,10 +53,11 @@ describe('Notification component', () => {
   it('marks all notifications as read', async () => {
     const notifications = [mock<Notification>({ messageRich: undefined })]
     const { wrapper, mocks } = getWrapper({ notifications })
+    await wrapper.vm.fetchNotificationsTask.perform()
     await wrapper.vm.fetchNotificationsTask.last
     await wrapper.find(selectors.markAll).trigger('click')
     expect(wrapper.find(selectors.notificationItem).exists()).toBeFalsy()
-    expect(mocks.$clientService.owncloudSdk.requests.ocs).toHaveBeenCalledTimes(2)
+    expect(mocks.$clientService.owncloudSdk.requests.ocs).toHaveBeenCalledTimes(3)
   })
   describe('avatar', () => {
     it('loads based on the username', async () => {
@@ -61,6 +66,7 @@ describe('Notification component', () => {
         user: 'einstein'
       })
       const { wrapper } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       const avatarImageStub = wrapper.findComponent<any>(selectors.avatarImageStub)
       expect(avatarImageStub.attributes('userid')).toEqual(notification.user)
@@ -74,6 +80,7 @@ describe('Notification component', () => {
         messageRichParameters: { user: { displayname, name } }
       })
       const { wrapper } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       const avatarImageStub = wrapper.findComponent<any>(selectors.avatarImageStub)
       expect(avatarImageStub.attributes('userid')).toEqual(name)
@@ -87,6 +94,7 @@ describe('Notification component', () => {
         message: undefined
       })
       const { wrapper } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       expect(wrapper.find(selectors.notificationSubject).exists()).toBeTruthy()
     })
@@ -98,6 +106,7 @@ describe('Notification component', () => {
         message: 'some message'
       })
       const { wrapper } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       wrapper.vm.showDrop()
       await wrapper.vm.$nextTick()
@@ -112,6 +121,7 @@ describe('Notification component', () => {
         }
       })
       const { wrapper } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       wrapper.vm.showDrop()
       await wrapper.vm.$nextTick()
@@ -127,6 +137,7 @@ describe('Notification component', () => {
         link: 'http://some-link.com'
       })
       const { wrapper } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       expect(wrapper.find(selectors.notificationLink).exists()).toBeTruthy()
     })
@@ -142,6 +153,7 @@ describe('Notification component', () => {
           }
         })
         const { wrapper } = getWrapper({ notifications: [notification] })
+        await wrapper.vm.fetchNotificationsTask.perform()
         await wrapper.vm.fetchNotificationsTask.last
         wrapper.vm.showDrop()
         await wrapper.vm.$nextTick()
@@ -168,6 +180,7 @@ describe('Notification component', () => {
           }
         })
         const { wrapper } = getWrapper({ notifications: [notification], spaces: [spaceMock] })
+        await wrapper.vm.fetchNotificationsTask.perform()
         await wrapper.vm.fetchNotificationsTask.last
         wrapper.vm.showDrop()
         await wrapper.vm.$nextTick()
@@ -187,6 +200,7 @@ describe('Notification component', () => {
         actions: [mock<NotificationAction>()]
       })
       const { wrapper } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       expect(wrapper.find(selectors.notificationActions).exists()).toBeTruthy()
     })
@@ -197,6 +211,7 @@ describe('Notification component', () => {
         actions: [mock<NotificationAction>({ link: 'http://some-link.com' })]
       })
       const { wrapper, mocks } = getWrapper({ notifications: [notification] })
+      await wrapper.vm.fetchNotificationsTask.perform()
       await wrapper.vm.fetchNotificationsTask.last
       expect(wrapper.find(selectors.notificationItem).exists()).toBeTruthy()
       const jsonResponse = {